Clean up GreyVest component library exports
