+++ src/greyVest/ButtonGroup.js
@@ -1,0 +1,4 @@
+import { defaultProps } from 'recompose'
+import Flex from '../layout/Flex'
+
+export default defaultProps({ className: 'gv-button-group' })(Flex)
