+++ src/greyVest/ErrorList.js
@@ -1,7 +1,21 @@
 import React from 'react'
 import _ from 'lodash/fp'
-import ErrorBlock from './ErrorBlock'
-import ErrorText from './ErrorText'
+import Flex from '../layout/Flex'
+
+let ErrorText = ({ children }) => (
+  <div className="gv-text-error">{children}</div>
+)
+
+let ErrorBlock = ({ children, ...props }) => (
+  <Flex className="gv-block-error" alignItems="center" {...props}>
+    <i className="material-icons" style={{ marginRight: 8 }}>
+      warning
+    </i>
+    <div>
+      <ErrorList>{children}</ErrorList>
+    </div>
+  </Flex>
+)
 
 let ErrorList = ({ block = false, children }) =>
   _.map(
@@ -13,4 +27,5 @@ let ErrorList = ({ block = false, children }) =>
       ),
     _.castArray(children)
   )
+
 export default ErrorList
