--- src/greyVest/ErrorText.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-// Error Text / List General Components
-let ErrorText = ({ children }) => (
-  <div className="gv-text-error">{children}</div>
-)
-
-export default ErrorText
