--- src/greyVest/FilterButtonList.js
@@ -1,8 +1,0 @@
-import BaseFilterButtonList from '../FilterButtonList'
-import React from 'react'
-
-let FilterButtonList = props => (
-  <BaseFilterButtonList {...props} className="gv-filter-button-list" />
-)
-
-export default FilterButtonList
