+++ src/greyVest/Modal.js
@@ -1,6 +1,4 @@
 import DefaultModal from '../layout/Modal'
 import { defaultProps } from 'recompose'
 
-let Modal = defaultProps({ className: 'gv-body' })(DefaultModal)
-
-export default Modal
+export default defaultProps({ className: 'gv-body' })(DefaultModal)
