+++ src/greyVest/index.js
@@ -1,37 +1,44 @@
-import { defaultProps } from 'recompose'
-import Flex from '../layout/Flex'
-
 export Box from './Box'
 export Button from './Button'
+export ButtonGroup from './ButtonGroup'
 export ButtonRadio from './ButtonRadio'
 export Checkbox from './Checkbox'
 export CheckboxList from './CheckboxList'
-export DateInput from './DateInput'
-export ErrorBlock from './ErrorBlock'
 export ErrorList from './ErrorList'
-export ErrorText from './ErrorText'
-export FilterButtonList from './FilterButtonList'
-export FilterListItem from './FilterListItem'
 export Fonts from './Fonts'
 export Icon from './Icon'
 export IconButton from './IconButton'
 export LinkButton from './LinkButton'
-export ListItem from './ListItem'
+export ListItem from './ListItem' // formerly also ListGroupItem
 export Modal from './Modal'
-export ModalPicker from './ModalPicker'
 export PagerItem from './PagerItem'
 export RadioList from './RadioList'
+export let SearchTree = () => {}
 export Select from './Select'
-export Style from './Style'
+export Style from './Style' // formerly GVStyle
 export { Tab, TabContent, TabLabel, Tabs } from './Tabs'
 export Table from './Table'
 export TabList from './TabList'
 export Textarea from './Textarea'
-export TextHighlight from './TextHighlight'
-export TextInput from './TextInput'
+export TextHighlight from './TextHighlight' // formerly Highlight
+export TextInput from './TextInput' // formerly Input
 export ToggleFiltersButton from './ToggleFiltersButton'
 export ToggleFiltersHeader from './ToggleFiltersHeader'
 export TreePauseButton from './TreePauseButton'
 
-export let SearchTree = () => {}
-export let ButtonGroup = defaultProps({ className: 'gv-button-group' })(Flex)
+// removed exports
+// ----------------
+// Adder (now generic FilterAdder)
+// CheckButton
+// Pager (now vanilla ResultPager exampleType)
+// PagedResultTable (now a new exampleType)
+// MissingTypeComponent (now theme only)
+// FilterList
+// AddableFilterList (now part of SearchFilters)
+// FiltersBox (now part of SearchFilters)
+// QueryBuilder
+// SearchLayout
+// SearchFilters (now a generic search interface)
+// FilterButtonList
+// StepsAccordion
+// QueryWizard
