+++ src/themes/greyVest/ModalPicker.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Flex from '../layout/Flex'
-import BaseModalPicker from '../layout/ModalPicker'
+import Flex from '../../layout/Flex'
+import BaseModalPicker from '../../layout/ModalPicker'
 
 let Label = (
   <Flex style={{ justifyContent: 'space-between', alignItems: 'center' }}>
