+++ src/themes/greyVest/index.js
@@ -4,14 +4,11 @@ import { defaultProps } from 'recompose'
 import Box from '../../greyVest/Box'
 import Button from '../../greyVest/Button'
 import Checkbox from '../../greyVest/Checkbox'
-import DateInput from '../../greyVest/DateInput'
-import FilterListItem from '../../greyVest/FilterListItem'
 import Fonts from '../../greyVest/Fonts'
 import Icon from '../../greyVest/Icon'
 import LinkButton from '../../greyVest/LinkButton'
 import ListItem from '../../greyVest/ListItem'
 import Modal from '../../greyVest/Modal'
-import ModalPicker from '../../greyVest/ModalPicker'
 import PagerItem from '../../greyVest/PagerItem'
 import RadioList from '../../greyVest/RadioList'
 import Select from '../../greyVest/Select'
@@ -21,7 +18,10 @@ import TextHighlight from '../../greyVest/TextHighlight'
 import TextInput from '../../greyVest/TextInput'
 
 //components used only for GreyVest theme
+import DateInput from './DateInput'
+import FilterListItem from './FilterListItem'
 import MissingTypeComponent from './MissingTypeComponent'
+import ModalPicker from './ModalPicker'
 import Tag from './Tag'
 
 export default {
